Map JSON Schema 'integer' type to MySQL BIGINTs rather than INTs
As Brion pointed out, with revision IDs already in the 600m range, we can
expect to need to log numbers outside the range of INTs (+/- 2.1b).

+++ server/eventlogging/jrm.py
@@ -84,7 +84,7 @@ COLUMN_DEFAULTS = {'type_': sqlalchemy.Unicode(STRING_MAX_LEN)}
 mappers = collections.OrderedDict((
     ('type', {
         'boolean': {'type_': sqlalchemy.Boolean},
-        'integer': {'type_': sqlalchemy.Integer},
+        'integer': {'type_': sqlalchemy.BigInteger},
         'number': {'type_': sqlalchemy.Float},
         'string': {'type_': sqlalchemy.Unicode(STRING_MAX_LEN)},
     }),
